+++ src/components/componentGroupButton/index.ts
@@ -1,4 +1,6 @@
 import { register } from '../../shared/functions';
 import { ComponentGroupButton } from './componentGroupButton';
 
+import '../componentEmoji/index';
+
 register('component-group-button', ComponentGroupButton);