+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # brote
 
+## 1.0.36
+
+### Patch Changes
+
+- Implementing new UI for group button
+
 ## 1.0.35
 
 ### Patch Changes