+++ src/components/componentGroupButton/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-group-button
 
+## 0.0.2
+
+### Patch Changes
+
+- Implementing new UI for group button
+
 ## 0.0.1
 
 ### Patch Changes